Fix compiler circular references
