+++ packages/compiler/src/config/types.ts
@@ -1,5 +1,5 @@
-import { Diagnostic, RuleRef } from "../core/index.js";
-import { YamlScript } from "../yaml/types.js";
+import type { Diagnostic, RuleRef } from "../core/index.js";
+import type { YamlScript } from "../yaml/types.js";
 
 /**
  * Represent the normalized user configuration.
