+++ packages/compiler/src/core/library.ts
@@ -1,6 +1,6 @@
 import { createDiagnosticCreator } from "./diagnostic-creator.js";
 import { compilerAssert } from "./diagnostics.js";
-import { Program } from "./program.js";
+import type { Program } from "./program.js";
 import { createJSONSchemaValidator } from "./schema-validator.js";
 import {
   DiagnosticMessages,
