+++ packages/compiler/src/testing/expect.ts
@@ -1,7 +1,7 @@
 import { fail, match, strictEqual } from "assert";
 import { Diagnostic, NoTarget, Type, formatDiagnostic, getSourceLocation } from "../core/index.js";
 import { isArray } from "../utils/misc.js";
-import { resolveVirtualPath } from "./test-host.js";
+import { resolveVirtualPath } from "./test-utils.js";
 
 /**
  * Assert there is no diagnostics.
