+++ packages/compiler/test/binder.test.ts
@@ -4,7 +4,7 @@ import { Binder, createBinder } from "../src/core/binder.js";
 import { createLogger } from "../src/core/logger/logger.js";
 import { createTracer } from "../src/core/logger/tracer.js";
 import { parse } from "../src/core/parser.js";
-import { Program } from "../src/core/program.js";
+import type { Program } from "../src/core/program.js";
 import { createSourceFile } from "../src/core/source-file.js";
 import {
   AliasStatementNode,
