+++ packages/compiler/test/checker/clone-type.test.ts
@@ -1,6 +1,6 @@
 import { deepStrictEqual, ok, strictEqual } from "assert";
 import { beforeEach, describe, it } from "vitest";
-import { Program } from "../../src/core/program.js";
+import type { Program } from "../../src/core/program.js";
 import { DecoratorContext, Type } from "../../src/core/types.js";
 import { TestHost, createTestHost } from "../../src/testing/index.js";
 import { createRekeyableMap } from "../../src/utils/misc.js";
