+++ packages/http/src/validate.ts
@@ -1,4 +1,4 @@
-import { Program } from "@typespec/compiler";
+import type { Program } from "@typespec/compiler";
 import { reportDiagnostic } from "./lib.js";
 import { getAllHttpServices } from "./operations.js";
 import { isSharedRoute } from "./route.js";
