+++ packages/playground/src/react/types.ts
@@ -1,4 +1,4 @@
-import { Program } from "@typespec/compiler";
+import type { Program } from "@typespec/compiler";
 import { ReactElement } from "react";
 
 export type CompilationCrashed = {
